Add rigid convected derivatives for Oldroyd
A slight modification to the convected derivatives allows for rigid structures,
here added for Oldroyd-A, Oldroyd-B, and Johnson-Segalman.

Note that I am not entirely sure about the equation for the viscoelastic stress
in this case. At this moment, they depend on the square of the conformation
tensor according to a small calculation based on the general thermodynamic-
consistent consitutive model.