Add Saramito model (first version)
We add the Saramito model, including also the Bingham model. The latter
is likely incorrect, with the former potentially having some issue. Note
that these are yield stress fluids, so a potential jump and change from
solid to liquid can give issues when solving flows for these models.

This is a first version, which may be corrected at a later point in
time. This version is added to have a base-line.