Add oscillatory shear tests
Adds SAOS (small-amplitude oscillatory shear) and LAOS (large-amplitude
oscillatory shear) tests.

For SAOS, we perform a simple direct simulation at a fixed frequency
together with a frequency sweep to obtain loss and storage moduli.

For LAOS, we obtain the Lissajous curves arranged in a Pipkin diagram.
This still misses the separation into the reversible (elastic) and
irreversible (viscous) part as well as the connected Chebyshev
polynomial decomposition. This will be added sometime in the future.